Added a folder for methods/classes that provide more general useful functions for testing, made appropriate changes to tests.
