+++ src/app/src/androidTest/java/ca/cmput301f21t22/nabu/intent/MyDayFragmentTest.java
@@ -28,6 +28,7 @@ import java.util.Calendar;
 
 import ca.cmput301f21t22.nabu.MainActivity;
 import ca.cmput301f21t22.nabu.R;
+import ca.cmput301f21t22.nabu.TestResources.DateProvider;
 
 /**
  * Runs tests on the My Day fragment.
@@ -36,44 +37,11 @@ import ca.cmput301f21t22.nabu.R;
  */
 public class MyDayFragmentTest {
     private Solo solo;
+    private DateProvider dateProvider;
     @Rule
     public ActivityTestRule<MainActivity> rule =
             new ActivityTestRule<>(MainActivity.class, true, true);
 
-    /**
-     * Creates a string representation of the current day
-     *
-     * @return day
-     *  A string representation of the current day
-     */
-    public String getCurrentDay(){
-        Calendar calendar = Calendar.getInstance();
-        int dayOfWeek = Calendar.DAY_OF_WEEK;
-        String day;
-        if (dayOfWeek == Calendar.SUNDAY) {
-            day = "Sun";
-        }
-        else if (dayOfWeek == Calendar.MONDAY) {
-            day = "Mon";
-        }
-        else if (dayOfWeek == Calendar.TUESDAY) {
-            day = "Tue";
-        }
-        else if (dayOfWeek == Calendar.WEDNESDAY) {
-            day = "Wed";
-        }
-        else if (dayOfWeek == Calendar.THURSDAY) {
-            day = "Thu";
-        }
-        else if (dayOfWeek == Calendar.FRIDAY) {
-            day = "Fri";
-        }
-        else {
-            day = "Sat";
-        }
-        return day;
-
-    }
     /**
      * Runs before all tests and creates solo instance.
      *
@@ -135,7 +103,7 @@ public class MyDayFragmentTest {
         solo.clickOnMenuItem("Habits");
         solo.clickOnView((FloatingActionButton) solo.getView(R.id.button_add_habit));
         solo.enterText((EditText) solo.getView(R.id.edit_title), "Farm Bees");
-        solo.clickOnToggleButton(this.getCurrentDay());
+        solo.clickOnToggleButton(this.dateProvider.getCurrentDay());
 
         solo.enterText((EditText) solo.getView(R.id.edit_reason), "I want honey");
         solo.clickOnText("Start Date");
@@ -196,7 +164,7 @@ public class MyDayFragmentTest {
         solo.clickOnMenuItem("Habits");
         solo.clickOnView((FloatingActionButton) solo.getView(R.id.button_add_habit));
         solo.enterText((EditText) solo.getView(R.id.edit_title), "Farm Bees");
-        solo.clickOnToggleButton(this.getCurrentDay());
+        solo.clickOnToggleButton(this.dateProvider.getCurrentDay());
 
         solo.enterText((EditText) solo.getView(R.id.edit_reason), "I want honey");
         solo.clickOnText("Start Date");
@@ -236,7 +204,7 @@ public class MyDayFragmentTest {
         solo.clickOnMenuItem("Habits");
         solo.clickOnView((FloatingActionButton) solo.getView(R.id.button_add_habit));
         solo.enterText((EditText) solo.getView(R.id.edit_title), "Farm Bees");
-        solo.clickOnToggleButton(this.getCurrentDay());
+        solo.clickOnToggleButton(this.dateProvider.getCurrentDay());
 
         solo.enterText((EditText) solo.getView(R.id.edit_reason), "I want honey");
         solo.clickOnText("Start Date");
